ml/requirements.txt
Flask-CORS
flask
scikit-learn
Pillow
numpy
scikit-image